LayoutInflater: Stop directly inflating TextViews
in fde8ebbf53c06db480448462c8ad17fdaffccf62, a try-catch was added in attempt to fix the out of bounds exception when directly inflating views, this is fine as a precautionary measure. However, the instantation of the views doesn't seem to throw exception and proceeds, causing the out of bounds, this might be caused by malformed/missing attributeset, to avoid SystemUI from crashing during light/dark theme transition, remove TextViews from the direct instanstation list.

Change-Id: I85cb6f5368b3b00bed8a222bc776cddf28765d3e