services: fix SystemUI/top-app CPU contention
during QS expansion, systemui is moved to top-app. this occurence results to 2 critical processes contending for CPU resources. If both SystemUI and the current top app have top-app scheduling group,
there may be CPU contention which causes animation jank. to fix this regression, we must demote the unseen process to a lower priority so it does not contest resources with systemui animation and avoid jank

Change-Id: I4ce856afad531abb2ef706aaaeb16fe4ac11ff20
Signed-off-by: rmp22 <195054967+rmp22@users.noreply.github.com>